4.21*: blogilistan laajennus, step10

+++ osa4/controllers/blogs.js
@@ -4,6 +4,7 @@ const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
 
+
 blogsRouter.get('/', async (request, response) => {
     let blogs = await Blog.find({}).populate('user', { username: 1, id: 1, name: 1})
      response.json(blogs)
@@ -12,31 +13,47 @@ blogsRouter.get('/', async (request, response) => {
   blogsRouter.post('/', async (request, response) => {
     try {
       let decodedToken = jwt.verify(request.token, process.env.SECRET)
+      console.log(decodedToken)
       if (!decodedToken.id) {
         return response.status(401).json({ error: 'token invalid' })
       }
+      let user = await User.findById(decodedToken.id)
+      const blog = new Blog(request.body)
+      blog.user = user._id
+      try {
+        let savedBlog = await blog.save()
+        user.blogs = user.blogs.concat(savedBlog._id)
+        await user.save()
+        response.status(201).json(savedBlog.toJSON())
+      } catch(exception) {
+        response.status(400).json(exception.message)
+      }
     } catch(exception) {
       return response.status(401).json({ error: 'token missing or invalid' })
-    }
-    let users = await User.findById(decodedToken.id)
-    let user = users[0]
-    const blog = new Blog(request.body)
-    blog.user = user._id
-    try {
-      let savedBlog = await blog.save()
-      user.blogs = user.blogs.concat(savedBlog._id)
-      await user.save()
-      response.status(201).json(savedBlog.toJSON())
-    } catch(exception) {
-      response.status(400).json(exception.message)
     }
   })
 
   blogsRouter.delete('/:id', async (request, response) => {
     try {
-      await Blog.findByIdAndRemove(request.params.id)
-      
+      let blog = await Blog.findById(request.params.id)
+      try {
+        let decodedToken = jwt.verify(request.token, process.env.SECRET)
+        if (!decodedToken.id) {
+          return response.status(401).json({ error: 'token invalid' })
+        }
+        if (blog.user.toString() === decodedToken.id.toString()) {
+          await Blog.findByIdAndRemove(request.params.id)
+          let user = await User.findById(decodedToken.id)
+          user.blogs = user.blogs.filter(blog => blog != request.params.id)
+          await user.save()
+        } else {
+          return response.status(401).json({error: 'Only creator can remove the blog'})
+        }
+      } catch(exception) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+      }
     } catch(exception) {
+      console.log(exception)
       //delete can always fake that things were done
     }
     response.status(204).end()
